requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
dill==0.3.8
distro==1.9.0
docopt==0.6.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.13.1
Flask==3.1.1
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2024.6.1
gensim==4.3.3
goto-statement==1.2
grip==4.6.2
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lightgbm==4.6.0
loguru==0.7.3
lxml==5.4.0
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
modelscope==1.28.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
nvitop==1.5.1
openai==1.97.1
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
parso==0.8.4
path-and-address==2.0.1
peft==0.7.1
pexpect==4.9.0
pillow==11.0.0
pip==25.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.13.1
seaborn==0.13.2
selenium==4.33.0
sentencepiece==0.2.0
setuptools==78.1.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
trio==0.30.0
trio-websocket==0.12.2
triton==3.2.0
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
